% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load.sample}
\alias{load.sample}
\title{Load a tiny simulation file}
\usage{
load.sample()
}
\value{
a data frame of a simulated data with $200$ sgRNAs
}
\description{
Load a tiny simulation file
}
\examples{
dat <- load.sample()
head(dat)
}
