% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load.sim}
\alias{load.sim}
\title{Load a simulation file from \href{https://github.com/hyunhwaj/Crispulator.jl}{hyunhwaj/Crispulator.jl}}
\usage{
load.sim(depth, facs, noise, effect)
}
\arguments{
\item{depth}{sequencing depth}

\item{facs}{bin ratio of FACS}

\item{noise}{noise ratio}

\item{effect}{effect size (proportion of negative/positive phenotype)}
}
\description{
Load a simulation file from \href{https://github.com/hyunhwaj/Crispulator.jl}{hyunhwaj/Crispulator.jl}
}
\examples{
library(CRISPRCloud2)
dat <- load.sim(facs=0.1, noise=1.0, effect=0.4)
df.ret <- mbetattest(X=dat, nci=4, na=4, nb=4, alpha=0.05)
}
