% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.mbttest}
\alias{run.mbttest}
\title{Run mbttest with given simulation data}
\usage{
run.mbttest(dat)
}
\arguments{
\item{data}{\code{data.frame} was generated by \href{crispulator.ucsf.edu}{CRISPulator}}
}
\value{
The two \code{data.frame} of statistics for each sgRNA and each gene
}
\description{
Run mbttest with given simulation data
}
\examples{
library(CRISPRCloud2)
dat <- read.csv("https://raw.githubusercontent.com/hyunhwaj/Crispulator.jl/master/simulation/matrix/scenario_4_0.1_0.5_0.2.csv")
run.mbttest(dat)
}
